requirements.txt
streamlit==0.71.0
pandas==1.1.4
numpy==1.19.4
Pillow==8.0.1
